+++ include/seqan3/utility/tuple_utility.hpp
@@ -14,8 +14,8 @@
 
 #include <utility>
 
-#include <seqan3/core/pod_tuple.hpp>
 #include <seqan3/utility/tuple/concept.hpp>
+#include <seqan3/utility/tuple/pod_tuple.hpp>
 #include <seqan3/utility/type_list/traits.hpp>
