+++ test/snippet/utility/tuple/pod_tuple.cpp
@@ -1,5 +1,5 @@
-#include <seqan3/core/pod_tuple.hpp>
 #include <seqan3/core/debug_stream.hpp>
+#include <seqan3/utility/tuple/pod_tuple.hpp>
 
 int main()
 {
