+++ test/unit/core/CMakeLists.txt
@@ -3,4 +3,3 @@ add_subdirectories()
 seqan3_test (add_enum_bitwise_operators_test.cpp)
 seqan3_test (bit_manipulation_test.cpp)
 seqan3_test (debug_stream_test.cpp)
-seqan3_test (pod_tuple_test.cpp)
