+++ test/unit/utility/tuple/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectories()
+
+seqan3_test (pod_tuple_test.cpp)
