+++ test/unit/utility/tuple_utility_test.cpp
@@ -12,12 +12,10 @@
 #include <gtest/gtest.h>
 
 #include <seqan3/core/detail/strong_type.hpp>
-#include <seqan3/core/pod_tuple.hpp>
+#include <seqan3/utility/tuple/pod_tuple.hpp>
 #include <seqan3/utility/tuple_utility.hpp>
 
 #include "my_tuple.hpp"
-
-
 struct bar : public seqan3::detail::strong_type<unsigned, bar>
 {
     using seqan3::detail::strong_type<unsigned, bar>::strong_type;
